Implement Interrupts, Link SDL2, and Bug Fixes.
Implement interrupt handler in the cpu, adding the interrupt
register in the bus and the master interrupt enable enable (lol)
to the cpu as the EI instruction takes an extra machine cycle
to take effect. The register in the bus is only for faster
read access rather than going through BusRead().

Link in SDL2 library with CMake and test initialization in
gbemu.c.

Fix DAA instruction to have implementation from emulator tutorial
found online. I'm not sure I fully understand the instruction
as my implementation was quite different from what was found
online, and that emulator was known to work (whereas mine is yet
to be tested).

Fix bug in Load instruction which was checking instruction
parameter 2 instead of 1.